Moved MuSig2 Round 1 out of the signing flow. Allow it without user confirmation.
The first round of MuSig does not involve access to the private keys and does not
complete authorizing a transaction. Therefore, it is safe to do it without user
confirmation, which allows software wallets to possibly do it in background.

In cases when the other cosigners are online, this allows to get a single-sig
user experience, as the user would only have to plug the device once, and confirm
a single action.