Increase MAX_N_KEYS_IN_WALLET_POLICY to 15.
This allows complex wallet policies with a larger number of cosigners.

The memory occupation increases by 160 bytes for devices using NBGL,
due to reserving a bigger number of UX pairs (stored in a global).

There is otherwise no increase in memory usage in SIGN_PSBT, and
the increase in REGISTER_WALLET is not problematic - plenty of memory
is available there.