Simplify signing flow removing MuSig2, and allow transactions with (#342)
only custom inputs.

The way the code was written, the signing flow would be skipped if
there is no internal inputs that can be signed with the wallet
policy. This is incorrect, as some apps might need to sign such
transactions.

This also removes the logic related to MuSig2, which is anyway
disabled and therefore redundant.