Harden crypto_derive_symmetric_key
The function would cause a buffer overflow passed a buffer longer
than 32 bytes. This is not an issue today as it was only ever used
for a fixed short string (WALLET_SLIP0021_LABEL).

The workaround to copy into a local buffer is still needed, so we
defensively forbid longer strings.